requirements.txt
botorch
pandas
flask
requests
python-dotenv
sqlalchemy
numpy
cma
icecream
openpyxl
firebase-admin
sqlite3
google-cloud